crypto: Add minimal support for cSHAKE and KMAC from NIST SP 800-185
In addition to SHA-3, NIST Special Publication 800-185 defines
additional "SHA-3 Derived Functions": cSHAKE, KMAC, TupleHash and
ParallelHash. KMAC is particularly useful as a simple replacement
for HMAC, to generate a Message Authentication Code or more generally
to generate a keyed-hash.

cSHAKE128 and cSHAKE256 are the base for all those functions. They are
identical to SHAKE128 and SHAKE256 except for a different padding.
Customized functions are built on top of cSHAKE by inputting (absorbing)
additional parameters with special encoding.

Add minimal support to produce cSHAKE and KMAC hashes by:

 - Extending the KecAcc accelerator (gem5 and XKCP backend) to support
   cSHAKE128 and cSHAKE256. It does not know about the cSHAKE encodings,
   it just absorbs/squeezes/pads data as for the SHA-3 variants.

 - Exposing cSHAKE128 and cSHAKE256 in the hash protocol on M3. Since
   the hash protocol does currently not support parameterized hash
   types, inputting the cSHAKE parameters is left up to the client side.
   This is simple, but has the disadvantage that clients may easily
   produce incompliant cSHAKE hashes by encoding the parameters
   incorrectly.

 - Adding an extra "cshake" crate that helps with encoding the cSHAKE
   and KMAC parameters.

 - Adding unit tests using the test vectors from the NIST website.

NOTE: The Rust backend of KecAcc (using the sha3 crate) does not support
cSHAKE yet. The cSHAKE support in the sha3 crate is on a higher level
where one would need to know the cSHAKE parameters in advance, so it
doesn't properly match the abstraction level of KecAcc.